- update typings + refactor
