--- src/enum/AlertType.ts
@@ -1,5 +1,0 @@
-export enum AlertType {
-  SUCCESS = 'success',
-  ERROR = 'error',
-  INFO = 'info',
-}
