+++ src/enum/Sorting.ts
@@ -3,3 +3,5 @@ export enum Sorting {
   ASC = 'asc',
   DESC = 'desc',
 }
+
+export default Sorting;
