--- src/enum/StatusVariant.ts
@@ -1,7 +1,0 @@
-export enum StatusVariant {
-  PRIMARY = 'primary',
-  ACCENT = 'accent',
-  SUCCESS = 'success',
-  ERROR = 'error',
-  DISABLED = 'disabled',
-}
