--- src/type/etc/FileResult.ts
@@ -1,1 +1,0 @@
-export type FileResult = { att: unknown[]; oldAtt: unknown[] };
