--- src/type/etc/MenuItem.ts
@@ -1,7 +1,0 @@
-import type { Nav } from './nav';
-
-export interface MenuItem extends Nav {
-  url?: string;
-  icon?: string;
-  onClick?: (key: string) => void;
-}
