--- src/type/etc/Rrange.ts
@@ -1,4 +1,0 @@
-export interface Range<T> {
-  min: T;
-  max: T;
-}
