--- src/type/etc/attachment.ts
@@ -1,5 +1,0 @@
-export interface Attachment {
-  id: string;
-  name: string;
-  url: string;
-}
