--- src/type/etc/file.ts
@@ -1,7 +1,0 @@
-export interface File {
-  created_at: string;
-  id: number;
-  name: string;
-  path: string;
-  updated_at: string;
-}
