--- src/type/etc/nav.ts
@@ -1,5 +1,0 @@
-export interface Nav {
-  title: string;
-  key: string;
-  isActive?: boolean;
-}
