--- src/type/etc/notification.ts
@@ -1,7 +1,0 @@
-export interface Notification {
-  id: string;
-  author: string;
-  date: string;
-  title: string;
-  description: string;
-}
