--- src/type/etc/position.ts
@@ -1,4 +1,0 @@
-export interface Position {
-  top: number | string;
-  left: number | string;
-}
