--- src/type/form/BadgeOption.ts
@@ -1,8 +1,0 @@
-import type { Option } from '../form/Option';
-
-export interface BadgeOption extends Option {
-  badge?: {
-    icon: string;
-    label: string;
-  };
-}
