--- src/type/form/DmpOption.ts
@@ -1,5 +1,0 @@
-import type { Option } from './Option';
-
-export interface DmpOption extends Option {
-  projectdescription?: string;
-}
