--- src/type/form/IconOption.ts
@@ -1,5 +1,0 @@
-import type { Option } from './Option';
-
-export interface IconOption extends Option {
-  icon?: string;
-}
