--- src/type/form/ImageOption.ts
@@ -1,5 +1,0 @@
-import type { Option } from './Option';
-
-export interface ImageOption extends Option {
-  image: string;
-}
