--- src/type/form/Option.ts
@@ -1,4 +1,0 @@
-export interface Option<T = string | number | boolean> {
-  label: string;
-  code: T;
-}
