--- src/type/form/ServiceOption.ts
@@ -1,5 +1,0 @@
-import type { Option } from './Option';
-
-export interface ServiceOption extends Option {
-  owner: Option;
-}
