--- src/type/notification/alert.ts
@@ -1,8 +1,0 @@
-import type { AlertType } from '../../enum/AlertType';
-
-export interface Alert {
-  id: string;
-  content: string;
-  type: AlertType;
-  ttl: number;
-}
