--- src/type/table/Thead.ts
@@ -1,8 +1,0 @@
-import { Sorting } from '../../enum/Sorting';
-
-export interface TheadProps {
-  label: string;
-  icon?: string;
-  sorting?: Sorting;
-  key: string;
-}
