+++ vite.config.ts
@@ -23,6 +23,7 @@ export default defineConfig({
       output: {
         globals: {
           vue: 'Vue',
+          lodash: 'lodash'
         },
       },
     },
